--- src/m3p2i_aip/config/physx.yml
@@ -1,5 +1,0 @@
-physics_engine: 'physx'
-num_threads: 8
-compute_device_id: 0
-graphics_device_id: 0
-use_gpu: True
